+++ src/sequences.cpp
@@ -11,7 +11,7 @@ using namespace Rcpp;
 // [[Rcpp::interfaces(r, cpp)]]
 // [[Rcpp::export]]
 NumericVector catalan(unsigned int n) {
-  std::vector<unsigned long long int> x(n);
+  std::vector<unsigned long int> x(n);
 
   x[0] = 1;
 
@@ -29,7 +29,7 @@ NumericVector catalan(unsigned int n) {
 // [[Rcpp::interfaces(r, cpp)]]
 // [[Rcpp::export]]
 NumericVector cullen(unsigned int n) {
-  std::vector<unsigned long long int> x(n);
+  std::vector<unsigned long int> x(n);
 
   x[0] = 0;
 
@@ -48,11 +48,11 @@ NumericVector cullen(unsigned int n) {
 // [[Rcpp::interfaces(r, cpp)]]
 // [[Rcpp::export]]
 NumericVector supercatalan(unsigned int n) {
-  std::vector<unsigned long long int> x(n + 1);
+  std::vector<unsigned long int> x(n + 1);
   x[0] = 1; x[1] = 1;
 
   double j = 3.0;
-  for (int i = 2; i <= x.size(); i++) {
+  for (unsigned int i = 2; i <= x.size(); i++) {
     x[i] = (3.0 * (2.0 * j - 3.0) * x[j - 2.0] - (j - 3.0) * x[j - 3.0]) / j;
     //x[i] = (1.0 / j) * ((6.0 * j - 9.0) * x[j - 1.0] - (j - 3.0) * x[j - 2]);
     j++;
@@ -82,7 +82,7 @@ NumericVector supercatalan(unsigned int n) {
 // [[Rcpp::interfaces(r, cpp)]]
 // [[Rcpp::export]]
 NumericVector fibonacci(unsigned long int n) {
-  std::vector<unsigned long long int> x(n);
+  std::vector<unsigned long int> x(n);
 
   x[0] = 1;
   x[1] = 1;
@@ -121,7 +121,7 @@ NumericVector fibonacci(unsigned long int n) {
 // [[Rcpp::interfaces(r, cpp)]]
 // [[Rcpp::export]]
 NumericVector three_n_one(unsigned int n) {
-  std::vector<unsigned long long int> x(n);
+  std::vector<unsigned long int> x(n);
 
   if (n == 1) {
     x.resize(1);
